Refactors files and folders to obtain an efficient folder structure
