+++ index.html
@@ -17,12 +17,9 @@
       defer
     ></script>
 
-    <script
-      type="module"
-      src="./src/js/api/listeners/auth/register.js"
-    ></script>
-    <script type="module" src="/src/js/api/index.js"></script>
-    <script type="module" src="/src/js/api/listeners/auth/login.js"></script>
+    <script type="module" src="./src/js/ui/listeners/auth/register.js"></script>
+    <script type="module" src="/src/js/index.js"></script>
+    <script type="module" src="/src/js/ui/listeners/auth/login.js"></script>
     <title>Auctio Home Page</title>
   </head>
   <body>
