+++ package.json
@@ -31,6 +31,7 @@
     "bootstrap-icons": "^1.11.3"
   },
   "devDependencies": {
+    "dotenv": "^16.4.5",
     "live-server": "^1.1.0",
     "npm-run-all": "^4.1.5",
     "sass": "^1.70.0"
