+++ src/js/api/auth/register.js
@@ -1,12 +1,10 @@
+import { authFetch } from "../autFetch.js";
 import { baseURL, registerEndpoint } from "../constants.js";
 
 export async function register(user) {
-  const response = await fetch(baseURL + registerEndpoint, {
+  const response = await authFetch(baseURL + registerEndpoint, {
     method: "post",
     body: JSON.stringify(user),
-    headers: {
-      "Content-Type": "application/json",
-    },
   });
   const json = await response.json();
   console.log(response);
