--- src/js/api/index.js
@@ -1,5 +1,0 @@
-import { fetchListings } from "./listings/listings.js";
-import { renderListings } from "./listings/renderListings.js";
-
-fetchListings();
-renderListings();
