+++ src/js/index.js
@@ -1,0 +1,5 @@
+import { fetchListings } from "./api/listings/fetchListings.js";
+import { renderListings } from "./api/listings/renderListings.js";
+
+fetchListings();
+renderListings();
