+++ src/js/storage/save.js
@@ -1,0 +1,3 @@
+export function save(key, value) {
+  localStorage(key, JSON.stringify(value));
+}
